src/test/java/daily/medium/LongestPalindromicSubstringTest.java
package daily.medium;

import org.junit.jupiter.api.Test;

import static org.junit.jupiter.api.Assertions.assertEquals;

class LongestPalindromicSubstringTest {

    LongestPalindromicSubstring longestPalindromicSubstring = new LongestPalindromicSubstring();

    @Test
    void longestPalindrome() {
        assertEquals("bab", longestPalindromicSubstring.longestPalindrome("babad"));
        assertEquals("bb", longestPalindromicSubstring.longestPalindrome("cbbd"));
    }

    @Test
    void longestPalindromeOwn() {
        assertEquals("bcb", longestPalindromicSubstring.longestPalindrome("bcbbd"));
    }
}